:twisted_rightwards_arrows: #104 - combine refactors into efforts for review page
